LPS-61720 Rename
