+++ modules/sdk/gradle-plugins-change-log-builder/samples/settings.gradle
