+++ modules/sdk/gradle-plugins-change-log-builder/src/main/resources/META-INF/gradle-plugins/com.liferay.change.log.builder.properties
@@ -1,0 +1,1 @@
+implementation-class=com.liferay.gradle.plugins.change.log.builder.ChangeLogBuilderPlugin
