--- modules/sdk/gradle-plugins-changelog-builder/samples/build.gradle
@@ -1,15 +1,0 @@
-buildscript {
-	dependencies {
-		classpath group: "com.liferay", name: "com.liferay.gradle.plugins.changelog.builder", version: "latest.release"
-	}
-
-	repositories {
-		mavenLocal()
-
-		maven {
-			url "http://cdn.repository.liferay.com/nexus/content/groups/public"
-		}
-	}
-}
-
-apply plugin: "com.liferay.changelog.builder"
