--- modules/sdk/gradle-plugins-changelog-builder/src/main/resources/META-INF/gradle-plugins/com.liferay.changelog.builder.properties
@@ -1,1 +1,0 @@
-implementation-class=com.liferay.gradle.plugins.changelog.builder.ChangeLogBuilderPlugin
